Updated entrypoint script and passing lnd port from docker socket if exists

+++ docker/entrypoint.sh
@@ -6,10 +6,22 @@ set -m
 #shuf -n 6 /relay/docker/unique-peer.txt >> /relay/.lnd/lnd.conf
 
 echo "Setting ENV vars..."
-export NODE_DOMAIN=$(curl $ECS_CONTAINER_METADATA_URI | echo $(jq -r .DockerName).$NODE_DOMAIN)
+
+if [ ! -z ${ECS_CONTAINER_METADATA_URI+x} ]; then
+  echo "ECS environment found, setting domain..."
+  export NODE_DOMAIN=$(curl $ECS_CONTAINER_METADATA_URI | echo $(jq -r .DockerName).$NODE_DOMAIN)
+fi
+
+# For dev purposes only
 # export NODE_DOMAIN=lvh.me
+
 export NODE_ALIAS=$HOSTNAME
 export NODE_IP=$NODE_SCHEME://$NODE_DOMAIN
+
+if [[ ! -f "/var/run/docker.sock" ]]; then
+  echo "Docker socket found, pulling LND port..."
+  export NODE_LND_PORT=$(curl --unix-socket /var/run/docker.sock http://localhost/containers/$HOSTNAME/json | jq '.NetworkSettings.Ports."9735/tcp"[0].HostPort')
+fi
 
 echo "Starting supervisor..."
 # Start the primary process and put it in the background
@@ -17,20 +29,24 @@ echo "Starting supervisor..."
 
 sleep 5
 
-echo "Creating wallet..."
-WALLET_PASSWORD=$(head /dev/urandom | tr -dc A-Za-z0-9 | head -c20)
-WALLET_PASSPHRASE=$(head /dev/urandom | tr -dc A-Za-z0-9 | head -c20)
+if [[ -f "/relay/.lnd/.lndpwd" ]]; then
+  echo "Wallet file found..."
+else
+  echo "Creating wallet..."
+  WALLET_PASSWORD=$(head /dev/urandom | tr -dc A-Za-z0-9 | head -c20)
+  WALLET_PASSPHRASE=$(head /dev/urandom | tr -dc A-Za-z0-9 | head -c20)
 
-echo $WALLET_PASSWORD >> /relay/.lnd/.lndpwd
-echo $WALLET_PASSPHRASE >> /relay/.lnd/.lndpwd
+  echo $WALLET_PASSWORD >> /relay/.lnd/.lndpwd
+  echo $WALLET_PASSPHRASE >> /relay/.lnd/.lndpwd
 
-expect /relay/docker/create_wallet.sh $WALLET_PASSWORD $WALLET_PASSPHRASE >> /relay/.lnd/.tempseed
+  expect /relay/docker/create_wallet.sh $WALLET_PASSWORD $WALLET_PASSPHRASE >> /relay/.lnd/.tempseed
 
-sed -e '2,14d' < /relay/.lnd/.tempseed >> /relay/.lnd/.lndseed
+  sed -e '2,14d' < /relay/.lnd/.tempseed >> /relay/.lnd/.lndseed
 
-rm /relay/.lnd/.tempseed
+  rm /relay/.lnd/.tempseed
 
-sleep 5
+  sleep 5
+fi
 
 echo "Starting relay..."
 supervisorctl start relay
